pyproject.toml
[project]
name = "personal-knowledge-ai"
version = "0.1.0"
description = "Personal knowledge AI assistant"
license = { text = "MIT AND (Apache-2.0 OR BSD-2-Clause)" }
requires-python = ">=3.11"
dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "asgiref==3.7.2",
  "bcrypt==4.3.0",
  "certifi==2025.4.26",
  "cffi==1.17.1",
  "click==8.1.8",
  "cryptography==45.0.3",
  "Django==5.0.2",
  "dnspython==2.7.0",
  "ecdsa==0.19.1",
  "email_validator==2.2.0",
  "fastapi==0.115.12",
  "fastapi-cli==0.0.7",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "numpy==1.26.4",
  "Jinja2==3.1.6",
  "jose==1.0.0",
  "markdown-it-py==3.0.0",
  "MarkupSafe==3.0.2",
  "mdurl==0.1.2",
  "mypy==1.15.0",
  "mypy_extensions==1.1.0",
  "passlib==1.7.4",
  "psycopg2-binary==2.9.10",
  "pyasn1==0.4.8",
  "pycparser==2.22",
  "pydantic==2.11.5",
  "pydantic-settings==2.9.1",
  "pydantic_core==2.33.2",
  "Pygments==2.19.1",
  "PyJWT==2.10.1",
  "python-dotenv==1.1.0",
  "python-jose==3.4.0",
  "python-multipart==0.0.20",
  "PyYAML==6.0.2",
  "python-frontmatter==1.1.0",
  "rich==14.0.0",
  "rich-toolkit==0.14.6",
  "rsa==4.9.1",
  "ruff==0.11.11",
  "shellingham==1.5.4",
  "six==1.17.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.41",
  "sqlparse==0.4.4",
  "starlette==0.46.2",
  "typer==0.15.4",
  "types-passlib==1.7.7.20250516",
  "types-pyasn1==0.6.0.20250516",
  "types-python-jose==3.4.0.20250516",
  "typing==3.7.4.3",
  "typing-inspection==0.4.1",
  "typing_extensions==4.13.2",
  "uvicorn==0.34.2",
  "uvloop==0.21.0",
  "watchdog==4.0.0",
  "watchfiles==1.0.5",
  "websockets==15.0.1",
]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "UP"]                      # Errors, Pyflakes, Import sorting, PyUpgrade
exclude = ["build", "dist", ".venv", "__pycache__"]


[tool.mypy]
strict_optional = true
python_version = "3.13"
strict = true
ignore_missing_imports = false
disallow_any_unimported = false
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true
